feature(Suggestions): auto-scroll to suggestion on arrow If the suggestions container is given a `max-height` & an `overflow-y: auto`, then the active suggestion outside of the viewable area will be hidden if selected via up/down arrow. While the suggestions list has an unbounded height by default, it seems likely that many would want to limit this in some way, as a very long suggestions list may look unpleasant in some scenarios, such as when the tags component is rendered inside of an element with a restricted height.
This change is not opt-in, but it should only have an effect if the list is height-limited in some way. So it seems appropriate to make this a default behavior. This way, the component functions as expected without having to hunt for a configuration option to set. But I understand if you want this to be opt-in anyway. Let me know @prakhar1989. This change can target v4.2.0.