fix(ReactTags): Drag and drop failing (#597)
* fix drag and drop not working

As mentioned in an issue, #559, if you drag and drop, React DND throws an invariant issue.

This change makes it so if you pass in a third property to each of your tags, you can have drag and drop. 

I haven't tested any other usecases beyond mine though, so if anyone else wanted to confirm this works for them, that would be appreciated. Meanwhile, the change can be made in your local `node_modules\react-tag-input\dist-modules\components\ReactTags.js#523`

* update Tag shape to include new key property

Co-authored-by: TankorSmash <tankorsmash+github@gmail.com>